Obs: Esqueci de dar o add//Finalização da autorização e autenticação de usuário, falta aplicar eles de maneira melhor; Toda a parte de autenticaçõ foi commentada; Resolvido o problema de reinicialização e fechamento sozinho;

+++ BackendContratanti/src/main/java/com/projeto/BackendContratanti/Dto/UsuarioResponseDTO.java
@@ -4,7 +4,7 @@ import com.projeto.BackendContratanti.Model.Usuario;
 
 import java.math.BigInteger;
 
-public record UsuarioResponseDTO(int id, String nome, String email , String telefone , String url_curriculo , String url_linkedin , String url_github , String cpf , String cidade , String senha , String descricao) {
+public record UsuarioResponseDTO(BigInteger id, String nome, String email , String telefone , String url_curriculo , String url_linkedin , String url_github , String cpf , String cidade , String senha , String descricao) {
     public UsuarioResponseDTO(Usuario usuario){
         this(usuario.getId(), usuario.getNome(), usuario.getEmail(), usuario.getTelefone(), usuario.getUrl_curriculo(), usuario.getUrl_linkedin(), usuario.getUrl_github(), usuario.getCpf(), usuario.getCidade(), usuario.getSenha(), usuario.getDescricao());
     }
